pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.121.2",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "playwright>=1.56.0",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
